web_app/requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.0
requests
pandas==1.5.3
plotly==5.15.0
gunicorn
Pillow==8.0.0
